Add `SKIP LOCKED` clause to DJ Pickup Query
This changes the optimized Postgres strategy to use the `SKIP LOCKED`
locking clause.

This is 'new' in Postgres 9.5, but thats pretty old for us. What it does
is change the locking behavior in this sub-query to do the pickup.
Before it would wait behind other locks, meaning that the more workers
were competing for jobs, the more locks were occuring.

This way new queries will just skip locked rows, and pick up the 'next'
available job!